# backend/docs/swagger.yaml
definitions:
...
  dto.AttachStudentRequest:
    properties:
      academic_year:
        description: '"2025-2026"'
        type: string
      semester:
        description: 1 или 2
        type: integer
      source:
        description: manual / sync
        type: string
      student_id:
        type: string
    type: object
  dto.AttachTeacherRequest:
    properties:
      teacher_id:
        type: string
    type: object
...
  dto.CreateDisciplineRequest:
    properties:
      code:
        type: string
      description:
        type: string
      external_id:
        type: string
      name:
        type: string
      source:
        description: manual / sync (default manual)
        type: string
    type: object
...
  dto.DisciplineResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      external_id:
        type: string
      id:
        type: string
      name:
        type: string
      source:
        type: string
      updated_at:
        type: string
    type: object
  dto.DisciplinesListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.DisciplineResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.UpdateDisciplineRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      birthday:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      group_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/disciplines:
    get:
      parameters:
      - description: Лимит (default 50)
        in: query
        name: limit
        type: integer
      - description: Смещение (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DisciplinesListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список дисциплин
      tags:
      - disciplines
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные дисциплины
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDisciplineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DisciplineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Code или Name уже заняты
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать дисциплину
      tags:
      - disciplines
  /api/disciplines/{id}:
    delete:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить дисциплину (soft-delete)
      tags:
      - disciplines
    get:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DisciplineResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Дисциплина по ID
      tags:
      - disciplines
    patch:
      consumes:
      - application/json
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDisciplineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DisciplineResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить дисциплину
      tags:
      - disciplines
  /api/disciplines/{id}/restore:
    post:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Восстановить дисциплину
      tags:
      - disciplines
  /api/disciplines/{id}/students:
    get:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Студенты дисциплины
      tags:
      - disciplines
    post:
      consumes:
      - application/json
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      - description: Студент и период обучения
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttachStudentRequest'
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Зачислить студента на дисциплину
      tags:
      - disciplines
  /api/disciplines/{id}/students/{user_id}:
    delete:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      - description: UUID студента
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отчислить студента с дисциплины
      tags:
      - disciplines
  /api/disciplines/{id}/teachers:
    get:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Преподаватели дисциплины
      tags:
      - disciplines
    post:
      consumes:
      - application/json
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      - description: ID преподавателя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttachTeacherRequest'
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Назначить преподавателя на дисциплину
      tags:
      - disciplines
  /api/disciplines/{id}/teachers/{user_id}:
    delete:
      parameters:
      - description: UUID дисциплины
        in: path
        name: id
        required: true
        type: string
      - description: UUID преподавателя
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Снять преподавателя с дисциплины
      tags:
      - disciplines
...
swagger: "2.0"